create list of video for video view
