+++ public/mocks/data.json
@@ -3,7 +3,7 @@
     "id": 1,
     "name": "MichelleVar",
     "title": "Esquiando en Dubai",
-    "videoSrc": "video3",
+    "videoSrc": "FXV-zwkRMxo",
     "avatarSrc": "imagenPerfil",
     "checkSrc": "Vector",
     "visitas": "1 M de visitas en 15 horas"
@@ -12,16 +12,16 @@
     "id": 2,
     "name": "Annete Black",
     "title": "Impressive Buildings",
-    "videoSrc": "impressive",
+    "videoSrc": "fNWxN0q-ZgE",
     "avatarSrc": "unsplash_XttWKETqCCQ",
     "checkSrc": "Vector",
     "visitas": "3 M de visitas en 10 dias"
   },
   {
     "id": 3,
-    "name": "Jefferson Vasquez",
+    "name": "Jefferson ",
     "title": "Animales salvajes",
-    "videoSrc": "video3",
+    "videoSrc": "fNWxN0q-ZgE",
     "avatarSrc": "perfil3",
     "checkSrc": "Vector",
     "visitas": "2K de visitas en 20 dias",
@@ -31,7 +31,7 @@
     "id": 4,
     "name": "Anderson Uribe",
     "title": "Planetas",
-    "videoSrc": "video4",
+    "videoSrc": "0pYOPvfarS8",
     "avatarSrc": "perfil4",
     "checkSrc": "Vector",
     "visitas": "2K de visitas en 20 dias",
@@ -41,7 +41,7 @@
     "id": 5,
     "name": "Andres Zuñiga",
     "title": "travel adventure",
-    "videoSrc": "video5",
+    "videoSrc": "jQ19FCO4MYA",
     "avatarSrc": "perfil5",
     "checkSrc": "Vector",
     "visitas": "3 M de visitas en 10 dias"
@@ -50,7 +50,7 @@
     "id": 6,
     "name": "Jefferson ",
     "title": "Otro Titulo",
-    "videoSrc": "video6",
+    "videoSrc": "6DNQWB3e0BU",
     "avatarSrc": "perfil6",
     "checkSrc": "Vector",
     "visitas": "2K de visitas en 20 dias",
@@ -60,7 +60,7 @@
     "id": 7,
     "name": "Andres Zuñiga",
     "title": "travel adventure",
-    "videoSrc": "video5",
+    "videoSrc": "afDXVnDnBf4",
     "avatarSrc": "perfil5",
     "checkSrc": "Vector",
     "visitas": "3 M de visitas en 10 dias"
@@ -69,16 +69,16 @@
     "id": 8,
     "name": "Annete Black",
     "title": "Impressive Buildings",
-    "videoSrc": "impressive",
+    "videoSrc": "2Lmz87uYBsw",
     "avatarSrc": "unsplash_XttWKETqCCQ",
     "checkSrc": "Vector",
     "visitas": "3 M de visitas en 10 dias"
   },
   {
     "id": 9,
-    "name": "Jefferson Vasquez",
+    "name": "Jefferson",
     "title": "Animales salvajes",
-    "videoSrc": "video3",
+    "videoSrc": "85L420nh4a0",
     "avatarSrc": "perfil3",
     "checkSrc": "Vector",
     "visitas": "2K de visitas en 20 dias",
