+++ src/index.scss
@@ -10,3 +10,4 @@
 @import './styles/components/VideoList.scss';
 @import './styles/components/Cards.scss';
 @import './styles/components/InputValidator.scss';
+@import './styles/pages/VideoShow.scss';
