+++ src/styles/pages/VideoShow.scss
@@ -5,3 +5,7 @@
 //area comentarios
 
 //area lista de videos
+.videoshow__list {
+  display: flex;
+  justify-content: flex-end;
+}
